Moved the module lib/slp.py into a package in lib/slp/
This will allow us to add various pieces such as the
slp_validator_0x1.py and other things in the future.
